Enable test harness for file tests
