+++ src/main/scala/Utils.scala
@@ -2,8 +2,19 @@ package fuselang
 
 object Utils {
 
-  def parseAst(s: String) = FuseParser.parse(s)
+  import scala.util.{Try, Success, Failure}
 
-  def typeCheck(s: String) = Compiler.compileString(s)
+  def parseAst(s: String) = Try(FuseParser.parse(s)) match {
+    case Success(p) => p
+    case Failure(Errors.ParserError(msg)) => throw new RuntimeException(s"$msg")
+    case Failure(f) => throw f
+  }
+
+  def typeCheck(s: String) = Try {
+    TypeChecker.checkFuse(FuseParser.parse(s))
+  } match {
+    case Success(p) => p
+    case Failure(f) => throw new RuntimeException(s"$f")
+  }
 
 }
