--- src/test/resources/should-compile/crlf.sea
@@ -1,5 +1,0 @@
-// This example tests support for Windows line endings.
-func add(a: float, b: float) {
-  let c = a + b;
-}
-
