--- src/test/resources/should-compile/matadd.sea
@@ -1,18 +1,0 @@
-/////////////////////////////////////////////
-// matrix matrix add
-// - This application uses two banked matrices as input,
-// one banked matrix as output, nested loops for each dimension,
-// regular banking in one dimension with each unroll
-// accessing different bank and all operands being matrices of
-// same parameters..
-////////////////////////////////////////////
-
-func madd(a: float[32][32 bank(32)], b: float[32][32 bank(32)], c: float[32][32 bank(32)]) {
-
-  for (let i = 0..31) {
-    for (let j = 0..31) unroll 32 {
-      c[i][j] := a[i][j] + b[i][j];
-    }
-  }
-
-}
