--- src/test/resources/should-compile/typedefs.sea
@@ -1,13 +1,0 @@
-type number = bit<32>;
-
-func add(a: number, b: number) {
-
-  // Allow integer addition between number types
-  let x = a+b;
-
-  // Allow integer addition between number and int types
-  let y = 5;
-  let z = a + b + x + y;
-
-}
-
