+++ src/test/resources/skip/matadd.sea
@@ -1,0 +1,8 @@
+func madd(a: float[32][32 bank(32)], b: float[32][32 bank(32)], c: float[32][32 bank(32)]) {
+  for (let i = 0..31) {
+    for (let j = 0..31) unroll 32 {
+      c[i][j] := a[i][j] + b[i][j];
+    }
+  }
+
+}
