+++ src/test/scala/TypePositive.scala
@@ -2,64 +2,9 @@ package fuselang
 
 import Utils._
 import Syntax._
+import org.scalatest.FunSuite
 
-class ParsingPositivea extends org.scalatest.FunSuite {
-  test("atoms parseAst") {
-    parseAst("1")
-    parseAst("true")
-    parseAst("true;")
-  }
-
-  test("binops") {
-    parseAst("1 + 2")
-    parseAst("1 + 2;")
-    parseAst("1 + 2 * 3;")
-    parseAst("1 + 2 * 3 >= 10 - 5 / 7;")
-    parseAst("true == false")
-  }
-
-  test("if") {
-    parseAst("if (true) {}")
-    parseAst("if (false) { 1 + 2 }")
-    parseAst("if (false) { 1 + 2 }")
-  }
-
-  test("decl") {
-    parseAst("decl x: bit<64>;")
-    parseAst("decl x: bool;")
-    parseAst("decl x: bool")
-    parseAst("decl x: bit<64>[10 bank 5];")
-  }
-
-  test("let") {
-    parseAst("let x = 1; x + 2;")
-  }
-
-  test("for loop") {
-    parseAst("""
-      for (let i = 0..10) {
-        x + 1;
-      }
-    """ )
-
-    parseAst("""
-      for (let i = 0..10) unroll 5 {
-        x + 1;
-      }
-    """ )
-  }
-
-  test("refresh banks") {
-    parseAst("""
-      x + 1;
-      ---
-      x + 2;
-    """ )
-  }
-
-}
-
-class TypeCorrect extends org.scalatest.FunSuite {
+class SimpleTypePositive extends FunSuite {
 
   test("Adding sized int to static int") {
     typeCheck("decl x: bit<64>; let y = 1; x + y;")
@@ -106,6 +51,22 @@ class TypeCorrect extends org.scalatest.FunSuite {
         let y = a[i];
       }
     """ )
+  }
+
+}
+
+class FileTypePositive extends FunSuite {
+
+  import java.nio.file.{Files, Paths}
+  import collection.JavaConverters._
+
+  val shouldCompile = Paths.get("src/test/resources/should-compile")
+
+  for (file <- Files.newDirectoryStream(shouldCompile).asScala) {
+    test(file.toString) {
+      val prog = new String(Files.readAllBytes(file))
+      typeCheck(prog)
+    }
   }
 
 }
